pyproject.toml
[project]
name = "smart-video-summarizer"
version = "0.1.0"
description = "AI-Powered Study Assistant for Students"
requires-python = ">=3.11"
dependencies = [
    # Web Framework
    "streamlit>=1.32.0",
    
    # AI/ML Libraries - Google
    "google-genai==1.49.0",
    
    # Keep these for document processing
    "pypdf2>=3.0.1",
    "python-docx>=1.1.0",
    "python-pptx>=0.6.23",
    "openpyxl>=3.1.2",
    
    # Audio/Video Processing
    "pydub>=0.25.1",
    
    # Vector Database (optional for now)
    "faiss-cpu>=1.7.4",
    
    # Utilities
    "python-dotenv>=1.0.1",
]



